clientApi: Reorganized nested modules and routes
Nested related modules within each other as follows:
- businesses
- photos
- profiles
    - follows
- posts
    - post_likes
    - post_photos
    - comments
        - comment_likes

Additionally implemented RouterModule to define nested routes.
