+++ client/src/posts/comments/comment_likes/models/comment_like.entity.ts
@@ -7,32 +7,30 @@ import {
   PrimaryColumn,
   UpdateDateColumn,
 } from 'typeorm';
-import { Profile } from '../../profiles/models/profile.entity';
+import { Comment } from '../../models/comment.entity';
+import { Profile } from '../../../../profiles/models/profile.entity';
 import { ApiProperty } from '@nestjs/swagger';
 
-@Entity('follows')
-export class Follow {
+@Entity('comment_likes')
+export class CommentLike {
+  @ApiProperty({ description: 'Liked comment id', example: 111 })
+  @PrimaryColumn()
+  comment_id: number;
+
   @ApiProperty({
-    description: 'Profile id of the user doing the following',
+    description: 'Profile id of the one performing a like',
     example: 1,
   })
   @PrimaryColumn()
-  follower: number;
+  profile_id: number;
 
-  @ApiProperty({
-    description: 'Profile id of the user being followed',
-    example: 2,
-  })
-  @PrimaryColumn()
-  followed: number;
+  @ManyToOne(() => Comment)
+  @JoinColumn({ name: 'comment_id' })
+  comment: Comment;
 
   @ManyToOne(() => Profile)
-  @JoinColumn({ name: 'follower' })
-  follower_profile: Profile;
-
-  @ManyToOne(() => Profile)
-  @JoinColumn({ name: 'followed' })
-  followed_profile: Profile;
+  @JoinColumn({ name: 'profile_id' })
+  profile: Profile;
 
   @CreateDateColumn({ type: 'timestamptz', default: () => 'CURRENT_TIMESTAMP' })
   created_at: Date;
