+++ client/src/posts/post_likes/models/post_like.entity.ts
@@ -7,27 +7,27 @@ import {
   PrimaryColumn,
   UpdateDateColumn,
 } from 'typeorm';
-import { Photo } from '../../photos/models/photo.entity';
-import { Post } from '../../posts/models/post.entity';
+import { Post } from '../../models/post.entity';
+import { Profile } from '../../../profiles/models/profile.entity';
 import { ApiProperty } from '@nestjs/swagger';
 
-@Entity('post_photos')
-export class PostPhoto {
-  @ApiProperty({ description: 'Photo id', example: 1 })
-  @PrimaryColumn()
-  photo_id: number;
-
-  @ApiProperty({ description: 'Post id', example: 111 })
+@Entity('post_likes')
+export class PostLike {
+  @ApiProperty({ description: 'Liked post id', example: 111 })
   @PrimaryColumn()
   post_id: number;
 
-  @ManyToOne(() => Photo)
-  @JoinColumn({ name: 'photo_id' })
-  photo: Photo;
+  @ApiProperty({ description: 'Profile id who performed the like', example: 1 })
+  @PrimaryColumn()
+  profile_id: number;
 
   @ManyToOne(() => Post)
   @JoinColumn({ name: 'post_id' })
-  profile: Post;
+  post: Post;
+
+  @ManyToOne(() => Profile)
+  @JoinColumn({ name: 'profile_id' })
+  profile: Profile;
 
   @CreateDateColumn({ type: 'timestamptz', default: () => 'CURRENT_TIMESTAMP' })
   created_at: Date;
